+++ src/main/java/ee/meriloo/toru/model/Point.java
@@ -1,28 +1,30 @@
 package ee.meriloo.toru.model;
 
+import java.math.BigDecimal;
+
 public class Point {
 
-    private Double xCoord;
-    private Double yCoord;
+    private BigDecimal xCoord;
+    private BigDecimal yCoord;
 
-    public Point(Double xCoord, Double yCoord) {
+    public Point(BigDecimal xCoord, BigDecimal yCoord) {
         this.xCoord = xCoord;
         this.yCoord = yCoord;
     }
 
-    public Double getxCoord() {
+    public BigDecimal getxCoord() {
         return xCoord;
     }
 
-    public void setxCoord(Double xCoord) {
+    public void setxCoord(BigDecimal xCoord) {
         this.xCoord = xCoord;
     }
 
-    public Double getyCoord() {
+    public BigDecimal getyCoord() {
         return yCoord;
     }
 
-    public void setyCoord(Double yCoord) {
+    public void setyCoord(BigDecimal yCoord) {
         this.yCoord = yCoord;
     }
 }
